pyproject.toml
[project]
name = "mcp-calendar"
version = "0.1.0"
description = "MCP tool for Google Calendar integration"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "mcp[cli]>=1.11.0",
    "google-auth-oauthlib>=1.2.0",
    "google-auth-httplib2>=0.2.0",
    "google-api-python-client>=2.0.0",
    "python-dotenv>=1.0.0",
    "python-dateutil>=2.8.2",
    "parsedatetime>=2.6",
]

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["src"]
[project.scripts]
mcp-calendar = "mcp_calendar.calendar:main"

